push

+++ src/main/java/dataStructure/SparseMatrix/MatrixElement.java
@@ -9,11 +9,11 @@ import java.io.Serializable;
  * @create: 2019-05-13 20:11
  */
 public class MatrixElement implements Serializable {
-    public int row;
-    public int col;
+    public long row;
+    public long col;
     public float val;
 
-    public void set(int row, int col,float val){
+    public void set(long row, long col,float val){
         this.col=col;
         this.row=row;
         this.val=val;
